[move/#229] 기타 파일들 이동
