+++ feature/home/src/main/java/com/napzak/market/home/HomeScreen.kt
@@ -46,8 +46,8 @@ import com.napzak.market.home.component.VerticalGridProducts
 import com.napzak.market.home.state.HomeUiState
 import com.napzak.market.product.model.Product
 import com.napzak.market.type.HomeBannerType
+import com.napzak.market.ui_util.ScreenPreview
 import com.napzak.market.ui_util.noRippleClickable
-import com.napzak.market.util.android.ScreenPreview
 import com.napzak.market.util.common.openUrl
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.ImmutableMap
